Move resolve_hostname_to_ipv4 to helpers.rs
